[CODE] Backup 255 changed file, end code changes, start Bugfixing

+++ Client/Data/Models/CooldownEventDto.cs
@@ -2,7 +2,7 @@
 {
     public class CooldownEventDto
     {
-        public ulong LastExecMs = 0;
+        public int LastExecMs = 0;
         public uint CooldownMs;
 
         public CooldownEventDto(uint cooldownMs)
