+++ Server/Data/Interfaces/ILoggingHandler.cs
@@ -18,6 +18,6 @@ namespace TDS_Server.Data.Interfaces
         void LogAdmin(LogType cmd, ITDSPlayer? source, string reason, int? targetid = null, bool asdonator = false, bool asvip = false, string? lengthOrEndTime = null);
         void LogKill(ITDSPlayer player, ITDSPlayer killer, uint weapon);
         void LogRest(LogType type, ITDSPlayer source, bool saveipserial = false, bool savelobby = false);
-        Task SaveTask(ulong? counter = null);
+        Task SaveTask(int? counter = null);
     }
 }
